+++ examples/example-remote/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 1.4.137
+
+### Patch Changes
+
+- Updated dependencies
+  - antd-management-fast-design-playground@1.0.144
+  - antd-management-fast-flow@1.0.89
+
 ## 1.4.136
 
 ### Patch Changes