+++ packages/antd-management-fast-flow/CHANGELOG.md
@@ -1,5 +1,11 @@
 # antd-management-fast-flow
 
+## 1.0.89
+
+### Patch Changes
+
+- update
+
 ## 1.0.88
 
 ### Patch Changes